+++ projects/igniteui-angular/src/lib/navigation-drawer/navigation-drawer.component.ts
@@ -204,6 +204,16 @@ export class IgxNavigationDrawerComponent implements
      */
     @Input() public width = '280px';
 
+
+    /**
+     * Enables/disables the animation, when toggling the drawer. Set to `false` by default.
+     * ````html
+     * <igx-nav-drawer [disableAnimation]="true"></igx-nav-drawer>
+     * ````
+     */
+    @HostBinding ('class.igx-nav-drawer--disable-animation')
+    @Input() public disableAnimation = false;
+
     /**
      * Width of the drawer in its mini state. Defaults to 68px.
      *